Improved PR and did some light refactoring of components
