+++ Content.Client/Storage/ItemCounterSystem.cs
@@ -1,4 +1,5 @@
 ﻿using Content.Shared.Storage;
+using Content.Shared.Storage.ItemCounter;
 using JetBrains.Annotations;
 using Robust.Shared.GameObjects;
