+++ Content.Server/Storage/Components/SharedItemCounter.cs
@@ -1,4 +1,5 @@
 ﻿using Content.Shared.Storage;
+using Content.Shared.Storage.ItemCounter;
 
 namespace Content.Server.Storage.Components
 {
