+++ Content.Server/Storage/ItemCounterSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Content.Server.Storage.Components;
 using Content.Shared.Storage;
+using Content.Shared.Storage.ItemCounter;
 using JetBrains.Annotations;
 using Robust.Shared.GameObjects;
