+++ Content.Shared/Storage/ItemCounter/SharedItemCounterComponent.cs
@@ -1,0 +1,10 @@
+﻿using Robust.Shared.GameObjects;
+
+namespace Content.Shared.Storage.ItemCounter
+{
+    [RegisterComponent]
+    public class SharedItemCounterComponent : Component
+    {
+        public override string Name => "ItemCounter";
+    }
+}
