--- Content.Shared/Storage/SharedStorageMapVisuals.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using Robust.Shared.Serialization;
-
-namespace Content.Shared.Storage
-{
-    [Serializable, NetSerializable]
-    public enum StorageMapVisuals : sbyte
-    {
-        AllLayers,
-        LayerChanged,
-    }
-}
